+++ cmd/gitbase/command/server_test.go
@@ -1,9 +1,15 @@
 package command
 
 import (
+	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
 	"testing"
 
+	fixtures "github.com/src-d/go-git-fixtures"
 	"github.com/stretchr/testify/require"
+	"gopkg.in/src-d/go-git.v4/plumbing"
 )
 
 func TestDirectories(t *testing.T) {
@@ -70,14 +76,14 @@ func TestDirectories(t *testing.T) {
 			path: "file:///siva/path?bare=true",
 			expected: directory{
 				Path: "/siva/path",
-				Bare: true,
+				Bare: bareOn,
 			},
 		},
 		{
 			path: "file:///siva/path?bare=false",
 			expected: directory{
 				Path: "/siva/path",
-				Bare: false,
+				Bare: bareOff,
 			},
 		},
 		{
@@ -118,7 +124,7 @@ func TestDirectories(t *testing.T) {
 			expected: directory{
 				Path:   "/siva/path",
 				Format: "git",
-				Bare:   false,
+				Bare:   bareOff,
 			},
 		},
 		{
@@ -146,4 +152,170 @@ func TestDirectories(t *testing.T) {
 			require.Equal(test.expected, dir)
 		})
 	}
+}
+
+func TestDiscoverBare(t *testing.T) {
+	defer func() {
+		require.NoError(t, fixtures.Clean())
+	}()
+
+	tmpDir, err := ioutil.TempDir("", "gitbase")
+	require.NoError(t, err)
+	defer os.RemoveAll(tmpDir)
+
+	emptyDir := filepath.Join(tmpDir, "empty")
+	err = os.Mkdir(emptyDir, 0777)
+	require.NoError(t, err)
+
+	bareDir := filepath.Join(tmpDir, "bare")
+	err = os.Mkdir(bareDir, 0777)
+	require.NoError(t, err)
+	dir := fixtures.ByTag("worktree").One().DotGit().Root()
+	err = os.Rename(dir, filepath.Join(bareDir, "repo"))
+	require.NoError(t, err)
+
+	nonBareDir := filepath.Join(tmpDir, "non_bare")
+	err = os.Mkdir(nonBareDir, 0777)
+	require.NoError(t, err)
+	dir = fixtures.ByTag("worktree").One().Worktree().Root()
+	err = os.Rename(dir, filepath.Join(nonBareDir, "repo"))
+	require.NoError(t, err)
+
+	tests := []struct {
+		path     string
+		bare     bareOpt
+		expected bool
+		err      bool
+	}{
+		{
+			path: "/does/not/exist",
+			err:  true,
+		},
+		{
+			path:     emptyDir,
+			bare:     bareAuto,
+			expected: false,
+		},
+		{
+			path:     emptyDir,
+			bare:     bareOn,
+			expected: true,
+		},
+		{
+			path:     emptyDir,
+			bare:     bareOff,
+			expected: false,
+		},
+		{
+			path:     bareDir,
+			bare:     bareAuto,
+			expected: true,
+		},
+		{
+			path:     bareDir,
+			bare:     bareOn,
+			expected: true,
+		},
+		{
+			path:     bareDir,
+			bare:     bareOff,
+			expected: false,
+		},
+		{
+			path:     nonBareDir,
+			bare:     bareAuto,
+			expected: false,
+		},
+		{
+			path:     nonBareDir,
+			bare:     bareOn,
+			expected: true,
+		},
+		{
+			path:     nonBareDir,
+			bare:     bareOff,
+			expected: false,
+		},
+	}
+
+	for _, test := range tests {
+		dir := directory{
+			Path: test.path,
+			Bare: test.bare,
+		}
+
+		t.Run(bareTestName(dir, test.err), func(t *testing.T) {
+			bare, err := discoverBare(dir)
+			if test.err {
+				require.Error(t, err)
+				return
+			}
+
+			require.NoError(t, err)
+			require.Equal(t, test.expected, bare)
+		})
+	}
+}
+
+func bareTestName(d directory, err bool) string {
+	bare := ""
+	switch d.Bare {
+	case bareOn:
+		bare = "bare"
+	case bareOff:
+		bare = "non-bare"
+	case bareAuto:
+		bare = "auto"
+	}
+
+	if err {
+		bare = "error"
+	}
+
+	return fmt.Sprintf("%s_%s", d.Path, bare)
+}
+
+func TestCache(t *testing.T) {
+	require := require.New(t)
+
+	tmpDir, err := ioutil.TempDir("", "gitbase")
+	require.NoError(err)
+	func() {
+		require.NoError(os.RemoveAll(tmpDir))
+	}()
+
+	server := &Server{
+		CacheSize:   512,
+		Format:      "siva",
+		Bucket:      0,
+		LogLevel:    "debug",
+		Directories: []string{"../../../_testdata"},
+		IndexDir:    tmpDir,
+	}
+
+	err = server.buildDatabase()
+	require.NoError(err)
+
+	cache := server.sharedCache
+	pool := server.pool
+	hash := plumbing.NewHash("dbfab055c70379219cbcf422f05316fdf4e1aed3")
+
+	_, ok := cache.Get(hash)
+	require.False(ok)
+
+	repo, err := pool.GetRepo("015da2f4-6d89-7ec8-5ac9-a38329ea875b")
+	require.NoError(err)
+
+	_, ok = repo.Cache().Get(hash)
+	require.False(ok)
+	require.Equal(cache, repo.Cache())
+
+	_, err = repo.CommitObject(hash)
+	require.NoError(err)
+
+	_, ok = cache.Get(hash)
+	require.True(ok)
+
+	_, ok = repo.Cache().Get(hash)
+	require.True(ok)
 }